Working on CurrentMonth frame
In CurrentMonth frame, the user can input the expenses of that month. This is displayed through a JTable, which is on the same frame. Users have the option to return to main menu. Working on incorporating the database next so that the program can load previous inputs as well as calculate the total expense